# api/swagger/v1/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hospitalApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Treatment": {
      "get": {
        "tags": [
          "Treatment"
        ],
        "operationId": "Treatment_GetAllTreatments",
        "parameters": [
          {
            "name": "FkPatientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Description",
            "in": "query",
            "schema": {
              "maxLength": 500,
              "minLength": 2,
              "type": "string"
            }
          },
          {
            "name": "Time",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.TreatmentOutputDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.TreatmentOutputDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.TreatmentOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Treatment"
        ],
        "operationId": "Treatment_Post",
        "parameters": [
          {
            "name": "staffId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.TreatmentInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.TreatmentInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.TreatmentInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/Treatment/{id}": {
      "get": {
        "tags": [
          "Treatment"
        ],
        "operationId": "Treatment_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.TreatmentOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.TreatmentOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.TreatmentOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Treatment"
        ],
        "operationId": "Treatment_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.TreatmentInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.TreatmentInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.TreatmentInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Treatment"
        ],
        "operationId": "Treatment_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "hospitalApi.DTOs.Inputs.TreatmentInputDto": {
        "type": "object",
        "properties": {
          "fkPatientId": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 500,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "hospitalApi.DTOs.Outputs.TreatmentOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fkPatientId": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 500,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}